add `duplicate_fastq_R1` option (#148)
Adds `duplicate_fastq_R1` option in `config.yaml`.

Can be set to "error", "warn" or "ignore" to specify what happens if the same FASTQ is used in multiple samples. Defaults to "error" if not specified.

This increments overall `dms-vep-pipeline-3` version to 3.16.0.

Addresses [this issue](https://github.com/dms-vep/dms-vep-pipeline-3/issues/147).